pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "spatialforge"
version = "0.1.0"
description = "Any Camera. Instant 3D. One API. — Spatial intelligence API platform powered by Depth Anything 3"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "SpatialForge" }]
keywords = ["3d", "depth-estimation", "spatial-intelligence", "computer-vision", "api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.18",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "redis>=5.2.0",
    "celery[redis]>=5.4.0",
    "minio>=7.2.0",
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "numpy>=1.26.0",
    "Pillow>=11.0.0",
    "opencv-python-headless>=4.10.0",
    "transformers>=4.47.0",
    "safetensors>=0.4.0",
    "httpx>=0.28.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "prometheus-client>=0.21.0",
    "structlog>=24.4.0",
    "orjson>=3.10.0",
]

[project.optional-dependencies]
tensorrt = ["tensorrt>=10.0.0"]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]
sdk = ["httpx>=0.28.0", "pydantic>=2.10.0"]

[project.scripts]
spatialforge = "spatialforge.main:run"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
